// src/app/models/credenciados/PutCredenciadosModel.ts
export class PutCredenciadosModel{
    idCredenciado: string = '';
    razaosocial: string = '';
    nomeCredenciado: string = '';
    cnpj: string = '';
    status: string = '';
    inscricaomunicipal: string = '';
    responsavel_sistema: string = '';
    email_usuario: string = '';
    senha_sistema: string = '';
    telefone_responsavel: string = '';
    idEndereco: string = '';

}